LayoutTests/js/typedarray-copyWithin-expected.txt

This test checks the behavior of the TypedArray.prototype.copyWithin function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.copyWithin.length is 2
PASS Int32Array.prototype.copyWithin.name is 'copyWithin'
PASS isSameFunctionForEachTypedArrayPrototype('copyWithin') is true
PASS testPrototypeReceivesArray('copyWithin', [undefined, this, { }, [ ], true, '']) is true
PASS testPrototypeFunction('copyWithin', '(0, 3)', [1, 2, 3, 4, 5], [4, 5, 3, 4, 5]) is true
PASS testPrototypeFunction('copyWithin', '(0, 3, 4)', [1, 2, 3, 4, 5], [4, 2, 3, 4, 5]) is true
PASS testPrototypeFunction('copyWithin', '(0, -2, -1)', [1, 2, 3, 4, 5], [4, 2, 3, 4, 5]) is true
PASS testPrototypeFunction('copyWithin', '(5, -5, 5)', [1, 2, 3, 4, 5], [1, 2, 3, 4, 5]) is true
PASS testPrototypeFunction('copyWithin', '(1, -5, 5)', [1, 2, 3, 4, 5], [1, 1, 2, 3, 4]) is true
PASS successfullyParsed is true

TEST COMPLETE
